requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==24.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
coverage==7.6.10
h11==0.14.0
idna==3.4
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==23.0
pluggy==1.0.0
PySocks==1.7.1
pytest==7.2.2
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==1.26.15
websocket-client==1.8.0
wsproto==1.2.0
